+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.5.0] - 2026-04-22
+
+### Added
+- `Schema#required_keys` — return the array of keys declared as required (including required nested schemas).
+- `Schema#optional_keys` — return the array of keys declared as optional.
+
 ## [0.4.0] - 2026-04-09
 
 ### Added